// Code/UBind.PortalApp/src/app/models/related-entity-type.enum.ts
// These enum determine related entity types.
export enum RelatedEntityType {
    Customer = "Customer",
    Quote = "Quote",
    Policy = "Policy",
    PolicyTransaction = "Policy Transaction",
    Claim = "Claim",
    Email = "Email",
    User = "User",
    Organisation = "Organisation",
    Portal = "Portal",
    Product = "Product",
    Release = "Release",
}